+++ packages/manager/cypress/support/ui/index.ts
@@ -1,4 +1,5 @@
 import * as actionMenu from './action-menu';
+import * as autocompletePopper from './autocomplete-popper';
 import * as breadcrumb from './breadcrumb';
 import * as buttons from './buttons';
 import * as dialog from './dialog';
@@ -13,6 +14,7 @@ import * as toggle from './toggle';
 
 export const ui = {
   ...actionMenu,
+  ...autocompletePopper,
   ...breadcrumb,
   ...buttons,
   ...dialog,