+++ packages/manager/src/features/Kubernetes/KubernetesClusterDetail/DeleteKubernetesClusterDialog.tsx
@@ -38,6 +38,12 @@ export const DeleteKubernetesClusterDialog = (props: Props) => {
   const disabled =
     preferences?.type_to_confirm !== false && confirmText !== clusterLabel;
 
+  React.useEffect(() => {
+    if (open && confirmText !== '') {
+      setConfirmText('');
+    }
+  }, [open]);
+
   const onDelete = () => {
     deleteCluster({ id: clusterId }).then(() => {
       onClose();