+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed:
+- `<RegionSelect />` can now dynamically get country flag and group all countrys #8996
+
+### Fixed:
+- Clear the Kubernetes Delete Dialog when it is re-opened #9000
+
+### Tech Stories:
+- MUIv5 Migration - Components > TagsInput, TagsPanel #8995
+
 ## [2023-04-03] - v1.90.0
 
 ### Added: